+++ src/app/kit/statistic/interfaces/chisquare.interface.ts
@@ -1,0 +1,8 @@
+import { RowStatsInterface } from '../models/row-stats.interface';
+
+export interface ChisquareInterface {
+	table?: RowStatsInterface[];
+	chisquare?: number;
+	percentile?: number;
+	pvalue?: number;
+}