+++ src/app/kit/statistic/models/col.stats.interface.ts
@@ -1,0 +1,7 @@
+export interface ColStatsInterface {
+	name?: string;
+	value?: number;
+	totCol?: number;
+	frequencyExpected?: number;
+	chiSquare?: number;
+}