+++ src/app/kit/statistic/models/row-stats.interface.ts
@@ -1,0 +1,7 @@
+import { ColStatsInterface } from './col.stats.interface';
+
+export interface RowStatsInterface {
+	name?: string;
+	values?: ColStatsInterface[];
+	totRow?: number;
+}