+++ src/app/kit/statistic/statistic.module.ts
@@ -1,0 +1,15 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+@NgModule({
+	declarations: [],
+	imports: [CommonModule],
+})
+export class StatisticModule {
+	static forRoot() {
+		return {
+			ngModule: StatisticModule,
+			providers: [],
+		};
+	}
+}